requirements.txt
PyQt6>=6.6.0
SQLAlchemy>=2.0.0
bcrypt>=4.0.0
psutil>=5.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pymysql>=1.1.0
cryptography>=41.0.0

